requirements.txt
# Core
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
scapy>=2.5.0
PyYAML>=6.0
joblib>=1.3

# Deep Learning
torch>=2.2

# Time-series storage
influxdb-client>=1.43.0

# Optional but helpful
psutil>=5.9
